SUBJ: High Scores Page MSG: 1) Implement a relatively crappy version of the final pages that shows high scores 2) Add the high score page to the flow of the game, following the end of game 3) Remove old score comparator, create proper and reliable implementation

+++ ScoreRecord.java
@@ -3,10 +3,10 @@
  */
 public class ScoreRecord {
     private final String name;
-    private final int score;
+    private final double score;
     private final String timestamp;
 
-    public ScoreRecord(String n, int s, String ts) {
+    public ScoreRecord(String n, double s, String ts) {
         this.name = n;
         this.score = s;
         this.timestamp = ts;
@@ -16,7 +16,7 @@ public class ScoreRecord {
         return this.name;
     }
 
-    public int getScore() {
+    public double getScore() {
         return this.score;
     }
